plugin-inbox: demote SyncTarget from Operation to a plain helper
SyncTarget's job is entirely plumbing — ensure a sync Routine exists,
then invoke its timer trigger via Trigger.TriggerMonitorService; the
trigger dispatcher is what actually invokes the connector's real sync
operation. That doesn't need Operation.make's registration, discovery,
or notify/spaceId-scoped-process machinery: it's not a domain action
in its own right, just a helper the graph action calls directly.

Moves it to util/sync-target.ts as a plain Effect-returning function,
resolving Trigger.TriggerMonitorService via Capabilities.ServiceResolver
(the same space-scoped layer resolution Operation.invoke uses under
the hood, just without spawning a tracked process) instead of
declaring it as an Operation service. Drops the sync-success/error
toasts along with it — the mailbox already surfaces feedback via its
progress meter, and the toolbar-only path never had toasts before
this action was unified across both surfaces.
