+++ packages/plugins/plugin-inbox/src/operations/index.ts
@@ -37,7 +37,6 @@ export const InboxOperationHandlerSet = OperationHandlerSet.lazy(
   () => import('./read-email'),
   () => import('./rename-filter'),
   () => import('./sync-contacts'),
-  () => import('./sync-target'),
   () => import('./unsubscribe-sender'),
   () => import('./sync-draft-events'),
 );
