+++ packages/plugins/plugin-inbox/src/util/index.ts
@@ -6,4 +6,5 @@ export * from './find-binding';
 export * from './mailbox-sync';
 export * from './match-filter';
 export * from './sync-routine';
+export * from './sync-target';
 export * from './util';
